Signup page frontend

+++ backend/opencity/src/main/java/it/unicam/cs/opencity/controller/UserController.java
@@ -3,8 +3,10 @@ package it.unicam.cs.opencity.controller;
 import it.unicam.cs.opencity.entity.User;
 import it.unicam.cs.opencity.service.RoleService;
 import it.unicam.cs.opencity.service.UserService;
+import it.unicam.cs.opencity.util.JwtTokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -15,11 +17,13 @@ public class UserController {
 
     private final UserService userService;
     private final RoleService roleService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Autowired
-    public UserController(UserService userService, RoleService roleService) {
+    public UserController(UserService userService, RoleService roleService, JwtTokenProvider jwtTokenProvider) {
         this.userService = userService;
         this.roleService = roleService;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @GetMapping("/")
@@ -37,25 +41,28 @@ public class UserController {
         }
     }
 
-    @GetMapping("/{id}/contents")
-    public ResponseEntity<Object> getUserContents(@PathVariable("id") String id) {
-
-        // TODO: implementare
-        return null;
-    }
-
-    @DeleteMapping("/remove/{id}")
-    public ResponseEntity<Object> removeUser(@PathVariable("id") String id) {
-        if(userService.removeUser(Integer.parseInt(id))) {
-            return ResponseEntity.ok("User successfully removed");
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Object> removeUser(@PathVariable("id") String id, @RequestHeader MultiValueMap<String, String> headers) {
+        if(hasPermission(headers, id)) {
+            if(userService.removeUser(Integer.parseInt(id))) {
+                return ResponseEntity.ok("User successfully removed");
+            } else {
+                return ResponseEntity.status(404).body("User not found");
+            }
         } else {
-            return ResponseEntity.status(404).body("User not found");
+            return ResponseEntity.status(403).body("Not authorized to remove this user");
         }
     }
 
     @GetMapping("/roles")
     public ResponseEntity<Object> getRoles() {
         return ResponseEntity.ok(roleService.getRoles());
+    }
+
+    private boolean hasPermission(MultiValueMap<String, String> headers, String userId) {
+        String token = headers.get("authorization").get(0).substring(7);
+        String id = jwtTokenProvider.extractId(token);
+        return id.equals(userId);
     }
 
 }
